Move *Requests forms to extrequests and extforms
